My first really cool project in python.
simulation of earth gravitational force. by changing the size we can also observe how the moon orbits earth. (ps) might have to make the win larger as of now it is now 800, 600. i also added a feature where it removes objects when it comes in contacts with earth of any of the borders. earth gravitational constant is 9.8, but this case we scaled it down and used 0.98. everything is not exactly right like the mass of the obj and the mass of the earth. it is all a visual simulation and is not meant to be used to predict anything. i also used math here. i use pygame and math. i used inverse tangent to find the angle theta and after that i used sin and cosinus to find all the other values of the accelration you will  see this use in line 47-50. also the formula to calculate distance between 2 points. which was used in line 44.